requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.29.1
google-api-core==2.5.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stop-words==2018.7.23
uritemplate==4.1.1
urllib3==1.26.8
wordcloud==1.8.1
